backapp/requirements.txt
Flask==1.1.2
google-api-core==1.22.4
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-cloud-bigquery==2.1.0
google-cloud-bigquery-storage==2.0.0
google-cloud-core==1.4.3
google-crc32c==1.0.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
gunicorn==20.0.4
numpy==1.16.0
pandas==0.24.2
pandas-gbq==0.14.0
pyarrow==1.0.1
requests==2.22.0
requests-html==0.10.0
requests-oauthlib==1.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
googletrans==3.0.0
google-cloud-translate==2.0.1
six==1.15.*
google-cloud-storage==1.35.0
gcsfs==0.7.1